File and Contents Reorganization
We reorganize the "documentation" folder and file. As discussed with
Rory Graves today, we are ghoing to separate two major parts.

One part, which here I call "description", will refer to the code and
the compiler "as it is", as a map or guide for reading the compiler
code, understanding the process and the algorithms, and where it is
based on. This part will serve those wishing to develope and improve
the compiler, but also those merely seeking to read it.

A separate part, which is the main intent behind this very project,
will describe the development and debugging process, and it will
cover issues about using git-github, writing tests, reporting bugs,
running and distributing the compiler, testing correctness and
performance, profiling, etc..

We also extract a "reference" file, to include a list of videos,
blog posts, or documents, that also shed light on scalac.